pyproject.toml
[project]
name = "tt-ticktick"
version = "0.1.0"
description = "Lightweight TickTick CLI for Claude Code"
authors = [
    { name = "Vaillus", email = "hugovaillaud@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "requests>=2.30",
    "python-dotenv>=1.0",
]

[project.scripts]
tt = "tt.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tt"]
